Route heroes render through ambient frame
